Remove explicit recentering of hull
In general the removed recentering would previously cause the
constructed hull to be missaligned with the THREE geometry. Only in the
cases when the geometry had symetric bounds about the origin or the
convex hull of the geometry does not contain the origin will the
constructed hull coincide with the THREE geometry.